+++ src/main/java/fr/miage/agents/api/message/negociation/ResultatNegociation.java
@@ -14,7 +14,7 @@ public class ResultatNegociation extends Message {
 
     public UUID session;
 
-    public int idProduit;
+    public long idProduit;
 
     public float prixNegocie;
